+++ tests/Unit/PulseAPIServiceProviderTest.php
@@ -3,8 +3,9 @@
 namespace Tests\Unit;
 
 use Robertogallea\PulseDBSize\Recorders\DBSizeRecorder;
+
 use function PHPUnit\Framework\assertNotNull;
 
 it('can load configuration', function () {
-    assertNotNull(config('pulse.recorders.') . DBSizeRecorder::class);
+    assertNotNull(config('pulse.recorders.').DBSizeRecorder::class);
 });