+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `pulse-db-size` will be documented in this file
 ## 1.0.1 - 2024-12-12
 
 - Add oracle support
+- Add configurable tables to ignore
 
 ## 1.0.0 - 2024-12-11